requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.8.2
jinja2==3.1.4
python-jose[cryptography]==3.3.0
# Base de datos
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1


